Move SlideTransition to types.tsx
